requirements.txt
praw>=7.7.0
nltk>=3.8.0
schedule>=1.2.0
python-dotenv>=1.0.0
requests>=2.31.0
